requirements.txt
fastapi
uvicorn[standard]
gunicorn
apscheduler
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
cryptography
twilio
google-api-python-client
google-auth
google-auth-oauthlib
requests
beautifulsoup4
python-multipart
google-genai
pydantic[email]
python-dotenv
